+++ apps/web/src/app/(afterLogin)/vocabulary/[id]/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React, { useState } from 'react'
-import { Button, Icon, Text } from '@vook-client/design-system'
+import { Button, Text } from '@vook-client/design-system'
 import { usePathname } from 'next/navigation'
 import {
   useDeleteBatchTermMutation,
@@ -9,19 +9,18 @@ import {
 } from '@vook-client/api'
 import { useQueryClient } from '@tanstack/react-query'
 
-import { Term } from '@/components/Term/Term'
 import { useModal } from '@/hooks/useModal'
 import { TermCreateModal, TermEditModal } from '@/modals/TermModal/TermModal'
 import { ModalTypes } from '@/hooks/useModal/useModal'
 import { useToast } from '@/hooks/useToast'
-
-import { warningContainer } from '../../layout.css'
+import { WarnBox } from '@/components/common/Common'
 
 import {
   vocabularyContainer,
   vocabularyHeader,
   vocabularyHeaderButton,
 } from './page.css'
+import { Term } from './_component/term/Term'
 
 export interface TermModalDataType {
   termUid: string
@@ -37,16 +36,17 @@ const Vocabulary = () => {
     meaning: '',
     synonym: [],
   })
-  const queryClient = useQueryClient()
-  const { addToast } = useToast()
-
   const [checkList, setCheckList] = useState<string[]>([])
-
   const [length, setLength] = useState(0)
-  const { open, toggleModal, type, setModal } = useModal()
+
+  const queryClient = useQueryClient()
   const path = usePathname()
-  const id = path.split('/').pop()
+  const { addToast } = useToast()
+  const { open, toggleModal, type, setModal } = useModal()
+
   const { data: response } = useVacabularyInfoQuery()
+  const id = path.split('/').pop()
+  const currentWorkspace = response?.result.find((value) => value.uid === id)
 
   const deleteBatchTermMutation = useDeleteBatchTermMutation(
     { termUids: checkList },
@@ -63,7 +63,6 @@ const Vocabulary = () => {
       },
     },
   )
-  const currentWorkspace = response?.result.find((value) => value.uid === id)
 
   const isDisabled = length >= 100
 
@@ -99,14 +98,9 @@ const Vocabulary = () => {
           </Button>
         </div>
       </div>
-      {isDisabled && (
-        <div className={warningContainer}>
-          <Icon name="alert-warning-big" />
-          <Text type="body-1" fontWeight="medium" color="label-neutral">
-            용어는 100개까지만 생성 가능합니다.
-          </Text>
-        </div>
-      )}
+
+      {isDisabled && <WarnBox>용어는 100개까지만 생성 가능합니다.</WarnBox>}
+
       <Term
         checkList={checkList}
         setCheckList={setCheckList}
@@ -114,6 +108,7 @@ const Vocabulary = () => {
         setModalData={setModalData}
         setLength={setLength}
       />
+
       {open && type === ModalTypes.CREATE && (
         <TermCreateModal uid={id as string} />
       )}
